Update commit icon in CommandPalette
This commit updates the icon used for the commit action within the CommandPalette component. Previously, it used the RewindIcon which seemed inappropriate for the action. The GitCommitIcon has been added to the icons index and used as a replacement. Additionally, an unnecessary class attribute was removed from the RewindIcon component.

Changes:
- Add GitCommitIcon to icons index
- Replace RewindIcon with GitCommitIcon in CommandPalette
- Remove unnecessary class attribute from RewindIcon