Enhance GUI with diff line count information
This change introduces a new feature to the user interface. It adds line count information to the GUI of each branch, displaying a summary of the lines added and removed. This is accomplished by modifying the rendering logic of each branch's card as well as counting and storing this information for future use.

Detail:
- In the rendering part, new div elements were added to hold added and removed information.
- In order to keep track of added and removed lines, a new method `sumBranchLinesAddedRemoved` was introduced that counts the number of lines added and removed in each branch by iterating over each line in the diff.
- Some styling changes were also made to accommodate these new components, such as modifying `flex-grow` to align the elements properly in the column.
- The required types were updated to handle the result of the new method