Added commit data to RemoteBranch struct
In this commit, a new field, "commits", of type Vec<VirtualBranchCommit>, was added to the RemoteBranch struct. This field is populated in the list_remote_branches function as part of the RemoteBranch creation process. Additionally, two assertions have been inserted into the test_detect_mergeable_branch function ensuring the "commits" collection is populated correctly.

Changes included:
- Append new "commits" field to RemoteBranch struct
- Updated the list_remote_branches function to populate commits
- Updated tests to include checks on the size of the "commits" array