Added branch notes functionality
The changes introduce a "notes" field for a branch. These notes can be set when a branch is created, updated with updating the branch, and read along with other details of the branch. This feature facilitates adding additional comments or description to a branch which are saved along the other metadata of the branch.

In detail:
- A "notes" field added to the Branch struct in various locations.
- "notes" added to the BranchUpdateRequest for updating.
- Saving and reading "notes" to/from disk is handled.
- A new action, 'updateBranchNotes', is added to the BranchController class for updating the notes field, along with corresponding error handling.
- The UI is updated to include a textarea for input of the notes, which triggers the 'updateBranchNotes' action on change