✨ feat: adds reset_files command to revert files
The "Revert File" function cheated a bit by undoing all the hunks, but that doesn't work for added or removed files. 

Added files are not removed, just contents cleared. Removed files are not restored.

This changes the implementation to just pass file paths and have git2 run the equivalent to `git checkout -- path` from the index.