Use Svelte context for file selection and ownership
This commit needlessly entangles two changes, but they are part of the same effort. We replace the use of props for passing selected files and selected ownership with context getters.

Landing this now since it's functionally much better than before, but this code needs more work. We are currently using a store on the `FileSelection` object essentially for retriggering checks, it has a smell to it.

Note that the reason selectedFileIds is separate from selectedFiles is that our file objects mutate with every update from the back end. That's why until now has been clearing quite often.

TODO: Rename `SelectedOwnership` to something more true to its function